requirements.txt
albumentations~=0.5.2
numpy~=1.19.3
segmentation-models-pytorch~=0.1.3
torch===1.7.1+cu110
-f https://download.pytorch.org/whl/torch_stable.html
Pillow~=8.1.0
Flask~=1.1.2
waitress~=1.4.4